+++ yolact_ros/msg/Segment.msg
@@ -4,5 +4,6 @@ int32 xmin
 int32 ymin
 int32 xmax
 int32 ymax
-int32[] encoded_pixels
+int32[] x_masks
+int32[] y_masks
 int32 pixel_size
