+++ yolact_ros/scripts/modules/bbox_recall.py
@@ -16,7 +16,7 @@ import random
 import numpy as np
 
 dump_file = 'weights/bboxes.pkl'
-aug_file  = 'weights/bboxes_aug.pkl'
+aug_file = 'weights/bboxes_aug.pkl'
 
 use_augmented_boxes = True
 
@@ -47,7 +47,7 @@ def jaccard(box_a, box_b, iscrowd=False):
     is simply the intersection over union of two boxes.  Here we operate on
     ground truth boxes and default boxes. If iscrowd=True, put the crowd in box_b.
     E.g.:
-        A ∩ B / A ∪ B = A ∩ B / (area(A) + area(B) - A ∩ B)
+        A  B / A  B = A  B / (area(A) + area(B) - A  B)
     Args:
         box_a: (tensor) Ground truth bounding boxes, Shape: [num_objects,4]
         box_b: (tensor) Prior boxes from priorbox layers, Shape: [num_priors,4]
@@ -55,16 +55,17 @@ def jaccard(box_a, box_b, iscrowd=False):
         jaccard overlap: (tensor) Shape: [box_a.size(0), box_b.size(0)]
     """
     inter = intersect(box_a, box_b)
-    area_a = ((box_a[:, 2]-box_a[:, 0]) *
-              (box_a[:, 3]-box_a[:, 1])).unsqueeze(1).expand_as(inter)  # [A,B]
-    area_b = ((box_b[:, 2]-box_b[:, 0]) *
-              (box_b[:, 3]-box_b[:, 1])).unsqueeze(0).expand_as(inter)  # [A,B]
+    area_a = ((box_a[:, 2] - box_a[:, 0]) *
+              (box_a[:, 3] - box_a[:, 1])).unsqueeze(1).expand_as(inter)  # [A,B]
+    area_b = ((box_b[:, 2] - box_b[:, 0]) *
+              (box_b[:, 3] - box_b[:, 1])).unsqueeze(0).expand_as(inter)  # [A,B]
     union = area_a + area_b - inter
 
     if iscrowd:
         return inter / area_a
     else:
         return inter / union  # [A,B]
+
 
 # Also convert to point form
 def to_relative(bboxes):
@@ -80,16 +81,17 @@ def make_priors(conv_size, scales, aspect_ratios):
     for j, i in product(range(conv_h), range(conv_w)):
         x = (i + 0.5) / conv_w
         y = (j + 0.5) / conv_h
-        
+
         for scale, ars in zip(scales, aspect_ratios):
             for ar in ars:
                 w = scale * ar / conv_w
                 h = scale / ar / conv_h
 
                 # Point form
-                prior_data += [x - w/2, y - h/2, x + w/2, y + h/2]
-    
+                prior_data += [x - w / 2, y - h / 2, x + w / 2, y + h / 2]
+
     return np.array(prior_data).reshape(-1, 4)
+
 
 # fixed_ssd_config
 # scales = [[3.5, 4.95], [3.6, 4.90], [3.3, 4.02], [2.7, 3.10], [2.1, 2.37], [2.1, 2.37], [1.8, 1.92]]
@@ -121,7 +123,7 @@ MEDIUM = 1
 LARGE = 2
 
 if __name__ == '__main__':
-        
+
     with open(dump_file, 'rb') as f:
         bboxes = pickle.load(f)
 
@@ -138,13 +140,12 @@ if __name__ == '__main__':
             sizes.append(LARGE)
 
     # Each box is in the form [im_w, im_h, pos_x, pos_y, size_x, size_y]
-    
+
     if use_augmented_boxes:
         with open(aug_file, 'rb') as f:
             bboxes_rel = pickle.load(f)
     else:
         bboxes_rel = to_relative(np.array(bboxes))
-        
 
     with torch.no_grad():
         sizes = torch.Tensor(sizes)
@@ -159,13 +160,12 @@ if __name__ == '__main__':
         chunk_size = 1000
         for i in range((bboxes_rel.size(0) // chunk_size) + 1):
             start = i * chunk_size
-            end   = min((i + 1) * chunk_size, bboxes_rel.size(0))
-            
+            end = min((i + 1) * chunk_size, bboxes_rel.size(0))
+
             ious = jaccard(bboxes_rel[start:end, :], anchors)
             maxes, maxidx = torch.max(ious, dim=1)
 
             perGTAnchorMax[start:end] = maxes
-    
 
         hits = (perGTAnchorMax > 0.5).float()
 
@@ -176,6 +176,3 @@ if __name__ == '__main__':
             _hits = hits[sizes == i]
             _size = (1 if _hits.size(0) == 0 else _hits.size(0))
             print(metric + ' recall: %.2f' % ((torch.sum(_hits) / _size) * 100))
-
-
-
