+++ yolact_ros/scripts/modules/data/coco.py
