+++ yolact_ros/scripts/modules/data/grid.npy
