+++ yolact_ros/scripts/modules/data/scripts/COCO.sh
