+++ yolact_ros/scripts/modules/data/scripts/COCO_test.sh
