+++ yolact_ros/scripts/modules/data/yolact_example_0.png
