+++ yolact_ros/scripts/modules/data/yolact_example_1.png
