+++ yolact_ros/scripts/modules/data/yolact_example_2.png
