+++ yolact_ros/scripts/modules/layers/__init__.py
@@ -1,2 +1,2 @@
 from .functions import *
-from .modules import *
+from .loss import *
