+++ yolact_ros/scripts/modules/layers/output_utils.py
@@ -2,15 +2,15 @@
 
 
 import torch
-import torch.nn as nn
+# import torch.nn as nn
 import torch.nn.functional as F
 import numpy as np
 import cv2
 
-from data import cfg, mask_type, MEANS, STD, activation_func
-from utils.augmentations import Resize
-from utils import timer
-from .box_utils import crop, sanitize_coordinates
+from modules.data import cfg, mask_type, MEANS, STD, activation_func
+from modules.utils.augmentations import Resize
+# from utils import timer
+from modules.layers.box_utils import crop, sanitize_coordinates
 
 def postprocess(det_output, w, h, batch_idx=0, interpolation_mode='bilinear',
                 visualize_lincomb=False, crop_masks=True, score_threshold=0):
