+++ yolact_ros/scripts/modules/unpack_statedict.py
@@ -9,8 +9,8 @@ print('Loading state dict...')
 state = torch.load(sys.argv[1])
 
 if not os.path.exists(sys.argv[2]):
-	os.mkdir(sys.argv[2])
+    os.mkdir(sys.argv[2])
 
 print('Saving stuff...')
 for key, val in state.items():
-	torch.save(val, sys.argv[2] + key)
+    torch.save(val, sys.argv[2] + key)
