+++ yolact_ros/scripts/modules/utils/__init__.py
@@ -1,0 +1,1 @@
+from .augmentations import SSDAugmentation, FastBaseTransform
