+++ yolact_ros/scripts/train.py
@@ -1,22 +1,18 @@
-from data import *
-from utils.augmentations import SSDAugmentation, BaseTransform
-from utils.functions import MovingAverage, SavePath
-from utils import timer
-from layers.modules import MultiBoxLoss
+from catkin_ws.src.yolact_ros.yolact_ros.scripts.module.data import *
+from catkin_ws.src.yolact_ros.yolact_ros.scripts.module.utils import SSDAugmentation, BaseTransform
+from catkin_ws.src.yolact_ros.yolact_ros.scripts.module.utils.functions import MovingAverage, SavePath
+from catkin_ws.src.yolact_ros.yolact_ros.scripts.module.utils import timer
+from catkin_ws.src.yolact_ros.yolact_ros.scripts.module.layers import MultiBoxLoss
 from yolact import Yolact
 import os
-import sys
 import time
 import math
-from pathlib import Path
 import torch
 from torch.autograd import Variable
 import torch.nn as nn
 import torch.optim as optim
 import torch.backends.cudnn as cudnn
-import torch.nn.init as init
 import torch.utils.data as data
-import numpy as np
 import argparse
 import datetime
