--- yolact_ros/utils/__init__.py
@@ -1,1 +1,0 @@
-from .augmentations import SSDAugmentation
